+++ src/main/java/com/lee/vrg/common/bo/UserLocationLogBo.java
@@ -7,9 +7,9 @@ import javax.validation.constraints.Pattern;
 import org.hibernate.validator.constraints.NotEmpty;
 
 public class UserLocationLogBo {
-	private Integer id;
+	private Long id;
 
-	private Integer userId;
+	private Long userId;
 
 	private String locationName;
 
@@ -23,19 +23,19 @@ public class UserLocationLogBo {
 
 	private Date gmtCreate;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
