+++ src/main/java/com/lee/vrg/common/model/LocationType.java
@@ -3,23 +3,23 @@ package com.lee.vrg.common.model;
 import java.util.Date;
 
 public class LocationType {
-    private Integer id;
+    private Long id;
 
     private String name;
 
     private String locationTypeDesc;
 
-    private Integer createrId;
+    private Long createrId;
 
     private Date gmtCreate;
 
     private Date gmtModify;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -39,11 +39,11 @@ public class LocationType {
         this.locationTypeDesc = locationTypeDesc == null ? null : locationTypeDesc.trim();
     }
 
-    public Integer getCreaterId() {
+    public Long getCreaterId() {
         return createrId;
     }
 
-    public void setCreaterId(Integer createrId) {
+    public void setCreaterId(Long createrId) {
         this.createrId = createrId;
     }
